rename dirs done
