+++ clean.sh
@@ -1,4 +1,4 @@
-rm -rf __pycache__/ coordinators/__pycache__ nodes/__pycache__ middleware/__pycache__ operations/__pycache__ joiners/__pycache__ workers/__pycache__ reducers/__pycache__
+rm -rf __pycache__/ src/coordinators/__pycache__ src/nodes/__pycache__ src/middleware/__pycache__ src/operations/__pycache__ src/joiners/__pycache__ src/workers/__pycache__ src/reducers/__pycache__
 
 rm -rf stats/
 rm -f pids.store
