+++ run-server.py
@@ -11,7 +11,7 @@ from subprocess import Popen
 
 PYTHON="python3"
 NODES_DIR="src/nodes/"
-CONFIG="--config=src/onfig.json"
+CONFIG="--config=src/config.json"
 
 def run(summary_workers,
         summary_reducers,
