+++ src/coordinators/replicator.py
@@ -5,10 +5,10 @@ from os import path
 
 sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 
-from src.middleware.connection import ReplicationSocket, GatherSocket
-from src.coordinators.stats import StatsManager
+from middleware.connection import ReplicationSocket, GatherSocket
+from coordinators.stats import StatsManager
 
-import src.middleware.constants as const
+import middleware.constants as const
 
 class DataReplicator(object):
