+++ src/nodes/local_points_filter.py
@@ -5,7 +5,7 @@ from os import path
 
 sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 
-from src.coordinators.filters import LocalPointsFilter
+from coordinators.filters import LocalPointsFilter
 
 def main(config):
