+++ src/nodes/local_team.py
@@ -5,7 +5,7 @@ from os import path
 
 sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 
-from src.joiners.count import LocalTeamCounter
+from joiners.count import LocalTeamCounter
 
 def main(workers, config):
