+++ src/reducers/match_summary.py
@@ -3,7 +3,7 @@ from os import path
 
 sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 
-from src.reducers.reducer import Reducer
+from reducers.reducer import Reducer
 
 class MatchSummaryReducer(Reducer):
