+++ src/workers/worker.py
@@ -3,7 +3,7 @@ from os import path
 
 sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 
-from src.middleware.connection import WorkerSocket
+from middleware.connection import WorkerSocket
 
 class Worker(object):
